pyproject.toml
[tool.poetry]
name = "epo"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
packages = [{ include = "epo", from = "src" }]

[tool.poetry.dependencies]
python = "^3.8.16"
pytorch-lightning = "^1.8.6"
transformers = "^4.25.1"
torchmetrics = "^0.11.0"
ray = {extras = ["tune"], version = "^2.2.0"}
protobuf = "3.20"
tqdm = "^4.64.1"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
isort = "^5.11.4"
mypy = "^0.991"
pytest = "^7.2.0"
types-tqdm = "^4.64.7.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
